move tracing to its own shared package because its not just about p2p
