+++ beacon-chain/node/BUILD.bazel
@@ -21,10 +21,11 @@ go_library(
         "//shared/cmd:go_default_library",
         "//shared/debug:go_default_library",
         "//shared/p2p:go_default_library",
+        "//shared/p2p/adapter:go_default_library",
         "//shared/p2p/adapter/metric:go_default_library",
-        "//shared/p2p/adapter/tracer:go_default_library",
         "//shared/params:go_default_library",
         "//shared/prometheus:go_default_library",
+        "//shared/tracing:go_default_library",
         "//shared/version:go_default_library",
         "@com_github_ethereum_go_ethereum//common:go_default_library",
         "@com_github_ethereum_go_ethereum//ethclient:go_default_library",
