--- shared/p2p/adapter/tracer/BUILD.bazel
@@ -1,21 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["tracer.go"],
-    importpath = "github.com/prysmaticlabs/prysm/shared/p2p/adapter/tracer",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//shared/p2p:go_default_library",
-        "@com_github_sirupsen_logrus//:go_default_library",
-        "@io_opencensus_go//exporter/jaeger:go_default_library",
-        "@io_opencensus_go//trace:go_default_library",
-    ],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["tracer_test.go"],
-    embed = [":go_default_library"],
-    deps = ["//shared/p2p:go_default_library"],
-)
