+++ shared/tracing/BUILD.bazel
@@ -1,0 +1,13 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["tracer.go"],
+    importpath = "github.com/prysmaticlabs/prysm/shared/tracing",
+    visibility = ["//visibility:public"],
+    deps = [
+        "@com_github_sirupsen_logrus//:go_default_library",
+        "@io_opencensus_go//exporter/jaeger:go_default_library",
+        "@io_opencensus_go//trace:go_default_library",
+    ],
+)
