+++ shared/tracing/README.md
